Fixes for various "crash on exit" bugs in 5L.
  * Fixed lots of bugs in TBTree, mostly in the code for removing nodes.
    TBTree should now work more or less correctly.
  * Removed the broken reference counting logic in TIndex and TIndexFile.
  * Made FatalError call abort(), not exit(1), so the destructors for
    (possibly corrupt) global variables will not be called.

This code may break either the Windows or Mac build; I'll try to fix things
right away.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@27297 b24aca02-f418-0410-9c49-835e4aa915c3

+++ Common/lang/old5l/TStream.cpp
@@ -657,7 +657,7 @@ TStream &FIVEL_NS operator>>(TStream &inStream, const ValueOrPercent &inVoP)
 	TStream arg_stream(argument.c_str());
 
 	// See if it appears to be a "(pcent ...)" command.
-	int prefix_len = strlen(PERCENT_COMMAND_PREFIX);
+	size_t prefix_len = strlen(PERCENT_COMMAND_PREFIX);
 	if (argument.length() > prefix_len &&
 		argument.substr(0, prefix_len) == PERCENT_COMMAND_PREFIX)
 	{
